Rename local analytics lib to tracker
To reduce confusion which the actual analaytics lib
